# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.5.0",
        "title": "Projeto Integrador SmartWater API",
        "description": "API criada a partir da biblioteca da Express, com o uso da linguagem Typescript, a API foi criada para obter dados de Caixas d'águas, Hidromêtros e Artesians emitidos por sensores",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/MIT"
        }
    },
    "host": "localhost:3000",
    "schemes": ["http", "https"],
    "basePath": "/api/timeseries/v0.5/smartcampusmaua/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
...
    "paths": {
        "/SmartLights": {
            "get" : {
                "tags": [
                    "SmartLights"
                ],
                "parameters": [
                    {
                        "name": "interval",
                        "in": "query",
                        "type": "number",
                        "description": "Query parameter that sets the interval of time in Minutes to get the data",
                        "default": "15",
                        "required": false
                    }
                ],
                "summary": "Get All Data emited by the Eight sensors installed for the SmartLights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "example": {
                                "SmartLight_1": {
                                    "data_temperature": [
                                        {
                                            "fieldValue": 11,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 1.277,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 2.08,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "...": ["..."]
                                },
                                "..." : {"..." : [{"...": "..."}]},
                                "SmartLight_8": {
                                    "rxInfo_latitude_0": [
                                        {
                                          "fieldValue": -23.64946,
                                          "timestamp": "2024-05-16T10:25:09.853963392Z",
                                          "start": "2024-05-16T10:22:43.901203384Z",
                                          "stop": "2024-05-16T22:22:43.901203384Z",
                                          "devEUI": 8066
                                        },
                                        {
                                          "fieldValue": -23.64946,
                                          "timestamp": "2024-05-16T10:25:16.850516604Z",
                                          "start": "2024-05-16T10:22:43.901203384Z",
                                          "stop": "2024-05-16T22:22:43.901203384Z",
                                          "devEUI": 8066
                                        }
                                    ],
                                    "fCnt": [
                                        {
                                          "fieldValue": 135,
                                          "timestamp": "2024-05-16T10:25:09.853963392Z",
                                          "start": "2024-05-16T10:22:43.901203384Z",
                                          "stop": "2024-05-16T22:22:43.901203384Z",
                                          "devEUI": 8066
                                        },
                                        {
                                          "fieldValue": 135,
                                          "timestamp": "2024-05-16T10:25:16.850516604Z",
                                          "start": "2024-05-16T10:22:43.901203384Z",
                                          "stop": "2024-05-16T22:22:43.901203384Z",
                                          "devEUI": 8066
                                        }
                                    ]
                                }
                                    
                                
                                
                            }
                        }
                    }
                }
            }
        },
        "/SmartLights/deviceName/{nodeName}": {
            "get" : {
                "tags": [
                    "SmartLights"
                ],
                "parameters": [
                    {
                        "name": "nodeName",
                        "in": "path",
                        "type": "string",
                        "description": "Param that receives the name of the SmartLight node, like SmartLight_1",
                        "required": true
                    }
                ],
                "summary": "Get All Data emited by the Eight sensors installed for the SmartLights",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "example": {
                                "SmartLight_1": {
                                    "data_temperature": [
                                        {
                                            "fieldValue": 11,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 1.277,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 2.08,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "...": ["..."]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "message": "The Route param nodeName is missing!"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "example": {
                                "message": "There is not a value with Nodename that you have passed!"
                            }
                        }
                    }
                }
            }
        },
        "/SmartLights/deviceId/{devEUI}": {
            "get" : {
                "tags": [
                    "SmartLights"
                ],
                "parameters": [
                    {
                        "name": "devEUI",
                        "in": "path",
                        "type": "string",
                        "description": "Param that receives the Id of the SmartLight node, 8066",
                        "required": true
                    }
                ],
                "summary": "Get Data emited filtering with the device ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "example": {
                                "SmartLight_1": {
                                    "data_temperature": [
                                        {
                                            "fieldValue": 11,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 1.277,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 2.08,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "...": ["..."]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "message": "The Route param devEUI is missing!"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "example": {
                                "message": "There is not a value with devEUI that you have passed!"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}